+++ python-magic/magic.py
@@ -14,17 +14,15 @@ Usage:
 'PDF document, version 1.2'
 >>>
 
-
 """
 
 import sys
 import glob
-import os.path
 import ctypes
 import ctypes.util
 import threading
 
-from ctypes import c_char_p, c_int, c_size_t, c_void_p
+from ctypes import c_char_p, c_int, c_size_t, c_void_p, byref, POINTER
 
 
 class MagicException(Exception):
@@ -36,11 +34,10 @@ class MagicException(Exception):
 class Magic:
     """
     Magic is a wrapper around the libmagic C library.
-
     """
 
     def __init__(self, mime=False, magic_file=None, mime_encoding=False,
-                 keep_going=False, uncompress=False):
+                 keep_going=False, uncompress=False, raw=False):
         """
         Create a new libmagic wrapper.
 
@@ -49,6 +46,7 @@ class Magic:
         magic_file - use a mime database other than the system default
         keep_going - don't stop at the first match, keep going
         uncompress - Try to look inside compressed files.
+        raw - Do not try to decode "non-printable" chars.
         """
         self.flags = MAGIC_NONE
         if mime:
@@ -57,13 +55,14 @@ class Magic:
             self.flags |= MAGIC_MIME_ENCODING
         if keep_going:
             self.flags |= MAGIC_CONTINUE
-
         if uncompress:
             self.flags |= MAGIC_COMPRESS
+        if raw:
+            self.flags |= MAGIC_RAW
 
         self.cookie = magic_open(self.flags)
         self.lock = threading.Lock()
-        
+
         magic_load(self.cookie, magic_file)
 
     def from_buffer(self, buf):
@@ -72,6 +71,11 @@ class Magic:
         """
         with self.lock:
             try:
+                # if we're on python3, convert buf to bytes
+                # otherwise this string is passed as wchar*
+                # which is not what libmagic expects
+                if type(buf) == str and str != bytes:
+                    buf = buf.encode('utf-8', errors='replace')
                 return maybe_decode(magic_buffer(self.cookie, buf))
             except MagicException as e:
                 return self._handle509Bug(e)
@@ -94,7 +98,13 @@ class Magic:
             return "application/octet-stream"
         else:
             raise e
-        
+
+    def setparam(self, param, val):
+        return magic_setparam(self.cookie, param, val)
+
+    def getparam(self, param):
+        return magic_getparam(self.cookie, param)
+
     def __del__(self):
         # no _thread_check here because there can be no other
         # references to this object at this point.
@@ -112,11 +122,13 @@ class Magic:
 
 _instances = {}
 
+
 def _get_magic_type(mime):
     i = _instances.get(mime)
     if i is None:
         i = _instances[mime] = Magic(mime=mime)
     return i
+
 
 def from_file(filename, mime=False):
     """"
@@ -129,6 +141,7 @@ def from_file(filename, mime=False):
     """
     m = _get_magic_type(mime)
     return m.from_file(filename)
+
 
 def from_buffer(buffer, mime=False):
     """
@@ -143,25 +156,25 @@ def from_buffer(buffer, mime=False):
     return m.from_buffer(buffer)
 
 
-
-
 libmagic = None
 # Let's try to find magic or magic1
-dll = ctypes.util.find_library('magic') or ctypes.util.find_library('magic1') or ctypes.util.find_library('cygmagic-1')
+dll = ctypes.util.find_library('magic') \
+    or ctypes.util.find_library('magic1') \
+    or ctypes.util.find_library('cygmagic-1')
 
-# This is necessary because find_library returns None if it doesn't find the library
+# necessary because find_library returns None if it doesn't find the library
 if dll:
     libmagic = ctypes.CDLL(dll)
 
 if not libmagic or not libmagic._name:
-    windows_dlls = ['magic1.dll','cygmagic-1.dll']
+    windows_dlls = ['magic1.dll', 'cygmagic-1.dll']
     platform_to_lib = {'darwin': ['/opt/local/lib/libmagic.dylib',
                                   '/usr/local/lib/libmagic.dylib'] +
-                         # Assumes there will only be one version installed
-                         glob.glob('/usr/local/Cellar/libmagic/*/lib/libmagic.dylib'),
+                       # Assumes there will only be one version installed
+                       glob.glob('/usr/local/Cellar/libmagic/*/lib/libmagic.dylib'),  # flake8:noqa
                        'win32': windows_dlls,
                        'cygwin': windows_dlls,
-                       'linux': ['libmagic.so.1'],    # fallback for some Linuxes (e.g. Alpine) where library search does not work
+                       'linux': ['libmagic.so.1'],  # fallback for some Linuxes (e.g. Alpine) where library search does not work # flake8:noqa
                       }
     platform = 'linux' if sys.platform.startswith('linux') else sys.platform
     for dll in platform_to_lib.get(platform, []):
@@ -177,12 +190,14 @@ if not libmagic or not libmagic._name:
 
 magic_t = ctypes.c_void_p
 
+
 def errorcheck_null(result, func, args):
     if result is None:
         err = magic_error(args[0])
         raise MagicException(err)
     else:
         return result
+
 
 def errorcheck_negative_one(result, func, args):
     if result is -1:
@@ -199,13 +214,14 @@ def maybe_decode(s):
         return s
     else:
         return s.decode('utf-8')
-    
+
+
 def coerce_filename(filename):
     if filename is None:
         return None
 
     # ctypes will implicitly convert unicode strings to bytes with
-    # .encode('ascii').  If you use the filesystem encoding 
+    # .encode('ascii').  If you use the filesystem encoding
     # then you'll get inconsistent behavior (crashes) depending on the user's
     # LANG environment variable
     is_unicode = (sys.version_info[0] <= 2 and
@@ -213,9 +229,10 @@ def coerce_filename(filename):
                   (sys.version_info[0] >= 3 and
                    isinstance(filename, str))
     if is_unicode:
-        return filename.encode('utf-8')
+        return filename.encode('utf-8', 'surrogateescape')
     else:
         return filename
+
 
 magic_open = libmagic.magic_open
 magic_open.restype = magic_t
@@ -238,6 +255,7 @@ _magic_file.restype = c_char_p
 _magic_file.argtypes = [magic_t, c_char_p]
 _magic_file.errcheck = errorcheck_null
 
+
 def magic_file(cookie, filename):
     return _magic_file(cookie, coerce_filename(filename))
 
@@ -245,6 +263,7 @@ _magic_buffer = libmagic.magic_buffer
 _magic_buffer.restype = c_char_p
 _magic_buffer.argtypes = [magic_t, c_void_p, c_size_t]
 _magic_buffer.errcheck = errorcheck_null
+
 
 def magic_buffer(cookie, buf):
     return _magic_buffer(cookie, buf, len(buf))
@@ -254,6 +273,7 @@ _magic_load = libmagic.magic_load
 _magic_load.restype = c_int
 _magic_load.argtypes = [magic_t, c_char_p]
 _magic_load.errcheck = errorcheck_negative_one
+
 
 def magic_load(cookie, filename):
     return _magic_load(cookie, coerce_filename(filename))
@@ -270,6 +290,26 @@ magic_compile = libmagic.magic_compile
 magic_compile.restype = c_int
 magic_compile.argtypes = [magic_t, c_char_p]
 
+_magic_setparam = libmagic.magic_setparam
+_magic_setparam.restype = c_int
+_magic_setparam.argtypes = [magic_t, c_int, POINTER(c_size_t)]
+_magic_setparam.errcheck = errorcheck_negative_one
+
+
+def magic_setparam(cookie, param, val):
+    v = c_size_t(val)
+    return _magic_setparam(cookie, param, byref(v))
+
+_magic_getparam = libmagic.magic_getparam
+_magic_getparam.restype = c_int
+_magic_getparam.argtypes = [magic_t, c_int, POINTER(c_size_t)]
+_magic_getparam.errcheck = errorcheck_negative_one
+
+
+def magic_getparam(cookie, param):
+    val = c_size_t()
+    _magic_getparam(cookie, param, byref(val))
+    return val.value
 
 
 MAGIC_NONE = 0x000000 # No flags
@@ -294,3 +334,11 @@ MAGIC_NO_CHECK_ASCII = 0x020000 # Don't check for ascii files
 MAGIC_NO_CHECK_TROFF = 0x040000 # Don't check ascii/troff
 MAGIC_NO_CHECK_FORTRAN = 0x080000 # Don't check ascii/fortran
 MAGIC_NO_CHECK_TOKENS = 0x100000 # Don't check ascii/tokens
+
+MAGIC_PARAM_INDIR_MAX = 0 # Recursion limit for indirect magic
+MAGIC_PARAM_NAME_MAX = 1 # Use count limit for name/use magic
+MAGIC_PARAM_ELF_PHNUM_MAX = 2 # Max ELF notes processed
+MAGIC_PARAM_ELF_SHNUM_MAX = 3 # Max ELF program sections processed
+MAGIC_PARAM_ELF_NOTES_MAX = 4 # # Max ELF sections processed
+MAGIC_PARAM_REGEX_MAX = 5 # Length limit for regex searches
+MAGIC_PARAM_BYTES_MAX = 6 # Max number of bytes to read from file