+++ python-omxplayer-wrapper/CHANGELOG.md
@@ -1,3 +1,19 @@
+# 0.3.0 -> 0.3.1
+* Remove `tests` package from distribution
+* Support calling `quit` multiple times without raising exception
+
+# 0.2.5 -> 0.3.0
+
+* Change `set_volume` to work with values between 0-10 instead of
+  millibels
+* Fix `volume` to return actual volume rather than just 1.0
+* Fix `rate` to return actual rate rather than just 1.0
+* Support providing arguments as a `str` which is then split with `shlex.split`,
+  i.e. you don't have to provide a list of shell split args if you don't want to.
+* Support `str` media file path in `OMXPlayer` constructor.
+* Cleanup omxplayer process on exit
+
+
 # 0.2.4 -> 0.2.5
 
 * Correct `omxplayer.__version__` to return 0.2.5 instead of 0.2.3 (in 0.2.4)